requirements.txt
absl-py==2.1.0
altgraph @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
astroid==2.15.7
asttokens==2.4.0
astunparse==1.6.3
bidict==0.23.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dill==0.3.7
dlib==19.24.4
docutils==0.20.1
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==24.3.25
fonttools==4.48.1
future @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
idna==3.7
importlib-resources==6.1.1
importlib_metadata==7.1.0
isort==5.12.0
jedi==0.19.0
keras==3.3.3
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
libclang==18.1.1
macholib @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
mtcnn==0.1.1
mypy==1.5.1
mypy-extensions==1.0.0
namex==0.0.8
numpy==1.26.3
opencv-python==4.10.0.82
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
parso==0.8.3
pillow==10.2.0
platformdirs==3.10.0
protobuf==4.25.3
pycrypto==2.6.1
Pygments==2.18.0
pylint==2.17.6
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
python-engineio==4.9.1
python-socketio==5.11.2
requests==2.32.3
rich==13.7.1
Send2Trash==1.8.2
simple-websocket==1.0.0
six @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
thonny==4.1.2
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
urllib3==2.2.1
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.15.0
wsproto==1.2.0
zipp==3.17.0
